fix: removed secrets resource block, output and variables. (#48)
* fix: removed secrets resource block, output and variables.

The keyvault module is only responsible for setting the keyvault and poicies. Content
setup is through other interfaces, like manual, api or app boostrapping.

* fix: added keyvault tags variable

The resource tags are not inherited from the resource group but set through variable.
Added the keyvault name to the output and normalize the output variable names to follow the terraform providers.

* fix: added terraform provider versions.

Both providers and terraform engine versions must be set by the module. 
Code was formatted as well.